requirements.txt
asyncio 
aiogram==2.25.2 
python-dotenv
